set default color back to grey
